Renames getstreamy to assemblit, renames components to _components
