+++ .gitignore
@@ -10,4 +10,5 @@ __pycache__
 .vscode
 .getstreamy
 .regression
+.assemblit
 testing.py
