+++ assemblit/__init__.py
@@ -1,0 +1,8 @@
+import assemblit.db
+import assemblit.vault
+import assemblit.web
+import assemblit.setup
+import assemblit.pages
+import assemblit.data_toolkit
+
+__all__ = ['db', 'vault', 'web', 'setup', 'pages', 'data_toolkit']
