+++ assemblit/_components/_selector.py
@@ -14,8 +14,8 @@ Contains the generic methods for a session-selector.
 
 import hashlib
 import streamlit as st
-from getstreamy import setup, db
-from getstreamy.components import _core, _key_value
+from assemblit import setup, db
+from assemblit._components import _core, _key_value
 from pytensils import utils
