+++ assemblit/pages/data_ingestion.py
@@ -15,8 +15,8 @@ Contains the `Class` for the data-ingestion-page.
 import copy
 import streamlit as st
 import pandas as pd
-from getstreamy import setup, db
-from getstreamy.components import _core, _data_uploader, _data_review
+from assemblit import setup, db
+from assemblit._components import _core, _data_uploader, _data_review
 
 
 class Content():
