+++ assemblit/pages/home.py
@@ -13,8 +13,8 @@ Contains the `Class` for the home-page.
 """
 
 import streamlit as st
-from getstreamy import setup, vault, web
-from getstreamy.components import _core
+from assemblit import setup, vault, web
+from assemblit._components import _core
 
 
 class Content():
