--- getstreamy/__init__.py
@@ -1,7 +1,0 @@
-import getstreamy.db
-import getstreamy.vault
-import getstreamy.web
-import getstreamy.setup
-import getstreamy.pages
-
-__all__ = ['db', 'vault', 'web', 'setup', 'pages']
